+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.14.0
+
+### Added
+
+- Read boolean values from config with the Config.get_bool method
+
+### Fixed
+
+- Save and send event after remove option
+
 ## 0.13.0
 
 ### Added